Fix: check if gdefault is set in mathmode.vim
Use a simple conditional to check whether the user sets `gdefault` in
`.vimrc` and adjust the substitute commands in `mathmode.vim` if
necessary.

Currently the function to check whether the cursor is pointing to a
location in math mode assumes that `nogdefault` is set (this is the
default behavior). If `gdefault` is set, then the `g` flag in each
occurrence of the substitute command will only count the first dollar or
double-dollar sign on each line, resulting in an incorrect count as well
as inappropriate math-mode expansion.

I've fixed this by wrapping the substitute calls in a conditional that
checks if `gdefault` is set and adjusts the flags accordingly.